Field of View Script
This add a field of viewscript that can be attached to an object to create a field of view using ray casting. The closest square is on another layer than the one that is for "Objects" This is to show that we can have the rays only penatrate certain things. You can also press q to move the field of view. It also takes input from another script if you want to see how I did that. Overall this should be a pretty robust system and should work for our  guards as we would just send a certain later of rays to check for player and if there is a hit we would trigger an alert status. LMK if yall have questions